pyproject.toml
[tool.poetry]
name = "es"
version = "0.1.0"
description = ""
authors = ["Łukasz Bacik <mail@luka.sh>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
elasticsearch = "~7.13.0"
requests = "^2.30.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
